feat(admin): rebuild admin UI with shadcn primitives + real data wiring
Dashboard
- Replace fake KPI trends and hardcoded activity rows with live data
  from ActivityRepo and a YoY revenue delta computed from RevenueRepo
- Extract RevenueChart to its own client component with a hover
  crosshair + tooltip showing month, current value, and prior-year value
- Top courses now use RevenueRepo.getTopCoursesByEnrollment which
  aggregates enrollment.priceAtPurchase per course (correct revenue,
  not current price * cumulative count)
- Drop the disabled "Coming soon" Calendar/Export buttons
- Fix int4 overflow: COALESCE(sum(price))::int now ::float8 in the
  monthly revenue and MTD revenue queries (priceAtPurchase is
  numeric(12,2), int4 caps around 2.1B THB)

Admin shell
- Mobile chip-strip nav replaced with a shadcn Sheet hamburger drawer
  that reuses the desktop NavLink (focus trap, scroll lock, Escape via
  Radix Dialog)
- Sidebar collapse state persists across navigations + reloads via
  localStorage, synchronized through useSyncExternalStore
- Static user card converted into a shadcn DropdownMenu with a
  destructive "Sign out" item

Lists
- /admin/users gains name/email search, role filter, pagination
- /admin/certificates gains code/name/course search, status filter,
  pagination; default ordering corrected to issuedAt desc

Slip queue
- AvatarInitials replaces the hardcoded #6366F1 -> #8B5CF6 gradient
- Redundant "ดู" action button removed (whole row was already clickable)

Courses
- Search submit now uses Button variant="primary" + shared <Input>
  instead of an inline accent-tinted button

+++ src/server/repos/admin-stats.ts
@@ -75,7 +75,7 @@ export const AdminStatsRepo = {
 				.where(and(eq(course.status, "published"), isNull(course.deletedAt))),
 			db
 				.select({
-					total: sql<number>`COALESCE(sum(${enrollment.priceAtPurchase})::int, 0)`,
+					total: sql<number>`COALESCE(sum(${enrollment.priceAtPurchase}), 0)::float8`,
 				})
 				.from(enrollment)
 				.where(
